Fix rendering issue
