+++ apps/builder/features/Builder/state/treeStore/hooks/useTreeSuspension.ts
@@ -3,6 +3,6 @@ import { TTreeContext } from "../TreeContext";
 
 export function useTreeSuspension(tree: TTreeContext["tree"]) {
   const {
-    nonSyncedStore: { synced: _synced },
+    nonSyncedStore: { synced },
   } = useSnapshot(tree);
 }
