+++ apps/builder/features/Data/useTreeId.ts
@@ -4,7 +4,7 @@ import { useRouter } from "next/router";
 export function useTreeId() {
   const { query } = useRouter();
 
-  if (!query.id || typeof query.id !== "string") {
+  if (query.id && typeof query.id !== "string") {
     throw new ODProgrammerError({
       code: "MISSING_URL_PARTS",
       message: `The url does not contain a valid tree id`,
