+++ apps/builder/pages/builder/[id]/index.tsx
@@ -5,12 +5,7 @@ import { ReactFlowProvider } from "react-flow-renderer";
 import { EditorHeader } from "../../../features/Builder/components/EditorHeader";
 import { SideMenu } from "../../../features/Builder/SideMenu";
 import { Layout } from "../../../components";
-import { LoadingSpinner, Stack } from "@open-decision/design-system";
 import { BuilderLayout } from "../../../features/Builder/components/BuilderLayout";
-
-export const getServerSideProps = async () => {
-  return { props: {} };
-};
 
 export default function BuilderPage() {
   return (
@@ -30,20 +25,12 @@ export default function BuilderPage() {
             }}
           />
           <SideMenu css={{ gridRow: "2", gridColumn: "1", layer: "1" }} />
-          <React.Suspense
-            fallback={
-              <Stack center>
-                <LoadingSpinner size="50px" />
-              </Stack>
-            }
-          >
-            <NodeEditor
-              css={{
-                gridColumn: "2 / 3",
-                gridRow: "2",
-              }}
-            />
-          </React.Suspense>
+          <NodeEditor
+            css={{
+              gridColumn: "2 / 3",
+              gridRow: "2",
+            }}
+          />
         </EditorProvider>
       </ReactFlowProvider>
     </Layout>
